requirements.txt
flask>=3.0
apscheduler>=3.10
requests>=2.31
python-dotenv>=1.0
